Refactor remote boost pad handling
